feat(deps): update imgui and implot dependencies
This commit updates the imgui dependency to v1.91.6 and implot to v0.16.
It also includes several other changes related to CMake configuration, documentation updates, and minor code adjustments across various modules.

Key changes include:
- **CMakeLists.txt**: Removed global GLM include, added explicit fmt dynamic linking for executables, and updated imgui/implot versions.
- **docs**: Updated `DATA.md` and `README.md` to reflect current status and progress. `TODO.md` status for chart rendering and static analysis fixes updated to "In Progress".
- **Backtester**: Minor adjustments in `backtester_engine.cpp` and `backtester_tab_controller.cpp` for type safety and clarity.
- **imgui_helpers.hpp**: Corrected `ImRect` calculation.
- **multi_timeframe_analyzer.h**: Added type alias for `MetricsCallback`.
- **workbench_core.cpp**: Changed `Toast` calls to use `ImGui::Toast`.
- **backend_tab_controller.cpp**: Corrected data type for equity curve plotting.
- **signals_tab_controller.cpp**: Corrected `AddRectFilled` parameters for crosshair rendering.
- **engine/CMakeLists.txt**: Removed global json_config.h include.
- **engine/glm_config.h**: Updated CUDA version definition and added GLM compiler definition for CUDA compatibility.

+++ src/apps/workbench/tabs/backend_tab_controller.cpp
@@ -217,8 +217,8 @@ void BackendTabController::renderBacktestingSuite() {
     ImGui::Text("Max Drawdown: %.2f", last_result_.max_drawdown);
     if (!equity_curve_.empty()) {
         if (ImPlot::BeginPlot("Equity Curve", ImVec2(-1,150))) {
-            std::vector<double> xs(equity_curve_.size());
-            for (size_t i = 0; i < xs.size(); ++i) xs[i] = static_cast<double>(i);
+            std::vector<float> xs(equity_curve_.size());
+            for (size_t i = 0; i < xs.size(); ++i) xs[i] = static_cast<float>(i);
             ImPlot::PlotLine("Equity", xs.data(), equity_curve_.data(), static_cast<int>(equity_curve_.size()));
             ImPlot::EndPlot();
         }
